requirements.txt
fastapi==0.120.0
openai-agents==0.4.2
sqlalchemy==2.0.44